Добавлен редактор статей с валидацией

+++ app/Http/Controllers/ArticleController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Http\Requests\ArticleFormRequest;
 use App\Models\Article;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Input;
 
 
 class ArticleController extends Controller
@@ -34,15 +35,14 @@ class ArticleController extends Controller
 
     public function delete($id)
     {
-        $author = \Auth::user()->name;
-        $check = Article::where('author', $author)->count();
+        $article = Article::find($id);
 
+        if (auth()->user()->name == $article->author) {
 
-        if ($check) {
             $delete__article = Article::find($id);
-        $delete__article ->delete();
-        return redirect()->route('home');
-            }
+            $delete__article ->delete();
+            return redirect()->route('home');
+        }
         else {
             return "У вас нет прав на удаление статьи";
         }
@@ -50,16 +50,26 @@ class ArticleController extends Controller
 
     public function edit($id)
     {
-        $author = \Auth::user()->name;
-        $check = Article::where('author', $author)->count();
+        $article = Article::find($id);
 
-
-        if ($check) {
-
-            return view('edit');
-        }
-        else {
+        if (auth()->user()->name == $article->author) {
+            return view('form',compact('article'));
+        } else {
             return "У вас нет прав на редактирование статьи";
         }
+    }
+
+
+    public function update(ArticleFormRequest $request, $id)
+    {
+        $article = Article::find($id);
+
+        $article -> title = Input::get('title');
+
+        $article -> text = Input::get('text');
+
+        $article -> save();
+
+        return redirect()->route('article', ['id' => $article->id]);
     }
 }
